Enum lldb_sys::LaunchFlags

pub enum LaunchFlags {
    None,
    Exec,
    Debug,
    StopAtEntry,
    DisableASLR,
    DisableSTDIO,
    LaunchInTTY,
    LaunchInShell,
    LaunchInSeparateProcessGroup,
    DontSetExitStatus,
    DetachOnError,
    ShellExpandArguments,
    CloseTTYOnExit,
}

Variants

None

Exec

Exec when launching and turn the calling process into a new process.

Debug

Stop as soon as the process launches to allow the process to be debugged.

StopAtEntry

Stop at the program entry point instead of auto-continuiing when launching or attaching at entry point.

DisableASLR

Disable address space layout randomization (ASLR).

DisableSTDIO

Disable stdio for the inferior process (e.g. for a GUI app).

LaunchInTTY

Launch the process in a new TTY if supported by the host.

LaunchInShell

Launch the process inside a shell to get shell expansion.

LaunchInSeparateProcessGroup

Launch the process in a separate process group.

DontSetExitStatus

If you are going to hand the process off (e.g. to debugserver), set this flag so that lldb and the handee don't race to set its exit status.

DetachOnError

If set, then the client stub should detach rather than killing the debugee.

ShellExpandArguments

Perform shell-style argument expansion.

CloseTTYOnExit

Close the open TTY on exit.