Crate lldb_sys

Enums

AccessType

AddressClass

BasicType

BreakpointEventType

ByteOrder
Byte order definitions.

CommandArgumentType

CommandFlags

ConnectionStatus

DescriptionLevel

DynamicValueType

EmulateInstructionOptions

Encoding
Register encoding defintions.

ErrorType

ExpressionEvaluationPhase

ExpressionResults

FilePermissions

Format
Display format definitions.

FrameComparison

FunctionNameType

GdbSignal

InputReaderAction

InputReaderGranularity

InstrumentationRuntimeType

LanguageType

LaunchFlags

MatchType

MemberFunctionKind

PathType

Permissions

QueueItemKind

QueueKind

RegisterKind

ReturnStatus

RunMode
Thread run modes.

SBAddressOpaque

SBAttachInfoOpaque

SBBlockOpaque

SBBreakpointLocationOpaque

SBBreakpointOpaque

SBBroadcasterOpaque

SBCommandInterpreterOpaque

SBCommandInterpreterRunOptionsOpaque

SBCommandOpaque

SBCommandPluginInterfaceOpaque

SBCommandReturnObjectOpaque

SBCommunicationOpaque

SBCompileUnitOpaque

SBDataOpaque

SBDebuggerOpaque

SBDeclarationOpaque

SBErrorOpaque

SBEventListOpaque

SBEventOpaque

SBExecutionContextOpaque

SBExpressionOptionsOpaque

SBFileSpecListOpaque

SBFileSpecOpaque

SBFrameOpaque

SBFunctionOpaque

SBHostOSOpaque

SBInputReaderOpaque

SBInstructionListOpaque

SBInstructionOpaque

SBLaunchInfoOpaque

SBLineEntryOpaque

SBListenerOpaque

SBModuleOpaque

SBModuleSpecListOpaque

SBModuleSpecOpaque

SBPlatformConnectOptionsOpaque

SBPlatformOpaque

SBPlatformShellCommandOpaque

SBProcessOpaque

SBQueueItemOpaque

SBQueueOpaque

SBSectionOpaque

SBSourceManagerOpaque

SBStreamOpaque

SBStringListOpaque

SBSymbolContextListOpaque

SBSymbolContextOpaque

SBSymbolOpaque

SBTargetRefOpaque

SBThreadCollectionOpaque

SBThreadPlanOpaque

SBThreadRefOpaque

SBTypeCategoryOpaque

SBTypeEnumMemberListOpaque

SBTypeEnumMemberOpaque

SBTypeFilterOpaque

SBTypeFormatOpaque

SBTypeListOpaque

SBTypeMemberFunctionOpaque

SBTypeMemberOpaque

SBTypeNameSpecifierOpaque

SBTypeOpaque

SBTypeSummaryOpaque

SBTypeSummaryOptionsOpaque

SBTypeSyntheticOpaque

SBUnixSignalsOpaque

SBValueListOpaque

SBValueOpaque

SBVariablesOptionsOpaque

SBWatchpointOpaque

ScriptLanguage

SectionType

StateType
Process and thread states.

StopReason
Thread stop reasons.

SymbolContextItem
These mask bits allow a common interface for queries that can limit the amount of information that gets parsed to only the information that is requested. These bits also can indicate what actually did get resolved during query function calls.

SymbolType

TemplateArgumentKind

TypeClass

TypeFlags

TypeOptions

TypeSummaryCapping

ValueType

WatchpointEventType

WatchpointKind

Constants

eFilePermissionsDirectoryDefault

eFilePermissionsFileDefault

eScriptLanguageDefault

Functions

CreateSBAddress

CreateSBAddress2

CreateSBAddress3

CreateSBAttachInfo

CreateSBAttachInfo2

CreateSBAttachInfo3

CreateSBBlock

CreateSBBreakpoint

CreateSBBreakpointLocation

CreateSBBroadcaster

CreateSBBroadcaster2

CreateSBCommand

CreateSBCommandInterpreterRunOptions

CreateSBCommandInterpreterRunOptions2

CreateSBCommandReturnObject

CreateSBCommunication

CreateSBCommunication2

CreateSBCompileUnit

CreateSBData

CreateSBDebugger

CreateSBDeclaration

CreateSBError

CreateSBEvent

CreateSBEvent2

CreateSBExecutionContext

CreateSBExecutionContext2

CreateSBExecutionContext3

CreateSBExecutionContext4

CreateSBExecutionContext5

CreateSBExpressionOptions

CreateSBFileSpec

CreateSBFileSpec2

CreateSBFileSpec3

CreateSBFileSpecList

CreateSBFrame

CreateSBFunction

CreateSBInputReader

CreateSBInstruction

CreateSBInstructionList

CreateSBLaunchInfo

CreateSBLineEntry

CreateSBListener

CreateSBListener2

CreateSBModule

CreateSBModule2

CreateSBModule3

CreateSBModuleSpec

CreateSBModuleSpecList

CreateSBPlatform

CreateSBPlatform2

CreateSBPlatformConnectOptions

CreateSBPlatformShellCommand

CreateSBProcess

CreateSBQueue

CreateSBQueueItem

CreateSBSection

CreateSBSourceManager

CreateSBSourceManager2

CreateSBStream

CreateSBStringList

CreateSBSymbol

CreateSBSymbolContext

CreateSBSymbolContextList

CreateSBTarget

CreateSBThread

CreateSBThreadCollection

CreateSBThreadPlan

CreateSBThreadPlan2

CreateSBType

CreateSBTypeCategory

CreateSBTypeEnumMember

CreateSBTypeEnumMemberList

CreateSBTypeFilter

CreateSBTypeFilter2

CreateSBTypeFormat

CreateSBTypeFormat2

CreateSBTypeFormat3

CreateSBTypeList

CreateSBTypeMember

CreateSBTypeMemberFunction

CreateSBTypeMemberFunction2

CreateSBTypeNameSpecifier

CreateSBTypeNameSpecifier2

CreateSBTypeNameSpecifier3

CreateSBTypeSummary

CreateSBTypeSummaryOptions

CreateSBTypeSynthetic

CreateSBUnixSignals

CreateSBValue

CreateSBValueList

CreateSBVariablesOptions

CreateSBVariablesOptions2

CreateSBWatchpoint

DisposeSBAddress

DisposeSBAttachInfo

DisposeSBBlock

DisposeSBBreakpoint

DisposeSBBreakpointLocation

DisposeSBBroadcaster

DisposeSBCommand

DisposeSBCommandInterpreter

DisposeSBCommandInterpreterRunOptions

DisposeSBCommandPluginInterface

DisposeSBCommandReturnObject

DisposeSBCommunication

DisposeSBCompileUnit

DisposeSBData

DisposeSBDebugger

DisposeSBDeclaration

DisposeSBError

DisposeSBEvent

DisposeSBExecutionContext

DisposeSBExpressionOptions

DisposeSBFileSpec

DisposeSBFileSpecList

DisposeSBFrame

DisposeSBFunction

DisposeSBInputReader

DisposeSBInstruction

DisposeSBInstructionList

DisposeSBLaunchInfo

DisposeSBLineEntry

DisposeSBListener

DisposeSBModule

DisposeSBModuleSpec

DisposeSBModuleSpecList

DisposeSBPlatform

DisposeSBPlatformConnectOptions

DisposeSBPlatformShellCommand

DisposeSBProcess

DisposeSBQueue

DisposeSBQueueItem

DisposeSBSection

DisposeSBSourceManager

DisposeSBStringList

DisposeSBSymbol

DisposeSBSymbolContext

DisposeSBSymbolContextList

DisposeSBTarget

DisposeSBThread

DisposeSBThreadCollection

DisposeSBThreadPlan

DisposeSBType

DisposeSBTypeCategory

DisposeSBTypeEnumMember

DisposeSBTypeEnumMemberList

DisposeSBTypeFilter

DisposeSBTypeFormat

DisposeSBTypeList

DisposeSBTypeMember

DisposeSBTypeMemberFunction

DisposeSBTypeNameSpecifier

DisposeSBTypeSummary

DisposeSBTypeSummaryOptions

DisposeSBTypeSynthetic

DisposeSBUnixSignals

DisposeSBValue

DisposeSBValueList

DisposeSBVariablesOptions

DisposeSBWatchpoint

SBAddressClear

SBAddressGetAddressClass

SBAddressGetBlock

SBAddressGetCompileUnit

SBAddressGetDescription

SBAddressGetFileAddress

SBAddressGetFunction

SBAddressGetLineEntry

SBAddressGetLoadAddress

SBAddressGetModule

SBAddressGetOffset

SBAddressGetSection

SBAddressGetSymbol

SBAddressGetSymbolContext

SBAddressIsValid

SBAddressOffsetAddress

SBAddressSetAddress

SBAddressSetLoadAddress

SBAttachInfoEffectiveGroupIDIsValid

SBAttachInfoEffectiveUserIDIsValid

SBAttachInfoGetEffectiveGroupID

SBAttachInfoGetEffectiveUserID

SBAttachInfoGetGroupID

SBAttachInfoGetIgnoreExisting

SBAttachInfoGetListener

SBAttachInfoGetParentProcessID

SBAttachInfoGetProcessID

SBAttachInfoGetProcessPluginName

SBAttachInfoGetResumeCount

SBAttachInfoGetUserID

SBAttachInfoGetWaitForLaunch

SBAttachInfoGroupIDIsValid

SBAttachInfoParentProcessIDIsValid

SBAttachInfoSetEffectiveGroupID

SBAttachInfoSetEffectiveUserID

SBAttachInfoSetExecutable

SBAttachInfoSetExecutable2

SBAttachInfoSetGroupID

SBAttachInfoSetIgnoreExisting

SBAttachInfoSetListener

SBAttachInfoSetParentProcessID

SBAttachInfoSetProcessID

SBAttachInfoSetProcessPluginName

SBAttachInfoSetResumeCount

SBAttachInfoSetUserID

SBAttachInfoSetWaitForLaunch

SBAttachInfoUserIDIsValid

SBBlockGetContainingInlinedBlock

SBBlockGetDescription

SBBlockGetFirstChild

SBBlockGetInlinedCallSiteColumn

SBBlockGetInlinedCallSiteFile

SBBlockGetInlinedCallSiteLine

SBBlockGetInlinedName

SBBlockGetNumRanges

SBBlockGetParent

SBBlockGetRangeEndAddress

SBBlockGetRangeIndexForBlockAddress

SBBlockGetRangeStartAddress

SBBlockGetSibling

SBBlockGetVariables

SBBlockGetVariables2

SBBlockIsInlined

SBBlockIsValid

SBBreakpointAddName

SBBreakpointClearAllBreakpointSites

SBBreakpointEventIsBreakpointEvent

SBBreakpointFindLocationByAddress

SBBreakpointFindLocationByID

SBBreakpointFindLocationIDByAddress

SBBreakpointGetBreakpointEventTypeFromEvent

SBBreakpointGetBreakpointFromEvent

SBBreakpointGetBreakpointLocationAtIndexFromEvent

SBBreakpointGetCondition

SBBreakpointGetDescription

SBBreakpointGetHitCount

SBBreakpointGetID

SBBreakpointGetIgnoreCount

SBBreakpointGetLocationAtIndex

SBBreakpointGetNames

SBBreakpointGetNumBreakpointLocationsFromEvent

SBBreakpointGetNumLocations

SBBreakpointGetNumResolvedLocations

SBBreakpointGetQueueName

SBBreakpointGetThreadID

SBBreakpointGetThreadIndex

SBBreakpointGetThreadName

SBBreakpointIsEnabled

SBBreakpointIsInternal

SBBreakpointIsOneShot

SBBreakpointIsValid

SBBreakpointLocationGetAddress

SBBreakpointLocationGetBreakpoint

SBBreakpointLocationGetCondition

SBBreakpointLocationGetDescription

SBBreakpointLocationGetID

SBBreakpointLocationGetIgnoreCount

SBBreakpointLocationGetLoadAddress

SBBreakpointLocationGetQueueName

SBBreakpointLocationGetThreadID

SBBreakpointLocationGetThreadIndex

SBBreakpointLocationGetThreadName

SBBreakpointLocationIsEnabled

SBBreakpointLocationIsResolved

SBBreakpointLocationIsValid

SBBreakpointLocationSetCondition

SBBreakpointLocationSetEnabled

SBBreakpointLocationSetIgnoreCount

SBBreakpointLocationSetQueueName

SBBreakpointLocationSetScriptCallbackBody

SBBreakpointLocationSetScriptCallbackFunction

SBBreakpointLocationSetThreadID

SBBreakpointLocationSetThreadIndex

SBBreakpointLocationSetThreadName

SBBreakpointMatchesName

SBBreakpointRemoveName

SBBreakpointSetCondition

SBBreakpointSetEnabled

SBBreakpointSetIgnoreCount

SBBreakpointSetOneShot

SBBreakpointSetQueueName

SBBreakpointSetScriptCallbackBody

SBBreakpointSetScriptCallbackFunction

SBBreakpointSetThreadID

SBBreakpointSetThreadIndex

SBBreakpointSetThreadName

SBBroadcasterAddInitialEventsToListener

SBBroadcasterAddListener

SBBroadcasterBroadcastEvent

SBBroadcasterBroadcastEventByType

SBBroadcasterClear

SBBroadcasterEventTypeHasListeners

SBBroadcasterGetName

SBBroadcasterIsValid

SBBroadcasterRemoveListener

SBCommandAddCommand

SBCommandAddMultiwordCommand

SBCommandGetHelp

SBCommandGetHelpLong

SBCommandGetName

SBCommandInterpreterAddCommand

SBCommandInterpreterAddMultiwordCommand

SBCommandInterpreterAliasExists

SBCommandInterpreterCommandExists

SBCommandInterpreterEventIsCommandInterpreterEvent

SBCommandInterpreterGetArgumentDescriptionAsCString

SBCommandInterpreterGetArgumentTypeAsCString

SBCommandInterpreterGetBroadcaster

SBCommandInterpreterGetBroadcasterClass

SBCommandInterpreterGetDebugger

SBCommandInterpreterGetIOHandlerControlSequence

SBCommandInterpreterGetProcess

SBCommandInterpreterGetPromptOnQuit

SBCommandInterpreterHandleCommand

SBCommandInterpreterHandleCommand2

SBCommandInterpreterHandleCommandsFromFile

SBCommandInterpreterHandleCompletion

SBCommandInterpreterHandleCompletion2

SBCommandInterpreterHasAliasOptions

SBCommandInterpreterHasAliases

SBCommandInterpreterHasCommands

SBCommandInterpreterIsActive

SBCommandInterpreterIsValid

SBCommandInterpreterRunOptionsGetAddToHistory

SBCommandInterpreterRunOptionsGetEchoCommands

SBCommandInterpreterRunOptionsGetPrintResults

SBCommandInterpreterRunOptionsGetStopOnContinue

SBCommandInterpreterRunOptionsGetStopOnCrash

SBCommandInterpreterRunOptionsGetStopOnError

SBCommandInterpreterRunOptionsSetAddToHistory

SBCommandInterpreterRunOptionsSetEchoCommands

SBCommandInterpreterRunOptionsSetPrintResults

SBCommandInterpreterRunOptionsSetStopOnContinue

SBCommandInterpreterRunOptionsSetStopOnCrash

SBCommandInterpreterRunOptionsSetStopOnError

SBCommandInterpreterSetPromptOnQuit

SBCommandInterpreterSourceInitFileInCurrentWorkingDirectory

SBCommandInterpreterSourceInitFileInHomeDirectory

SBCommandIsValid

SBCommandPluginInterfaceDoExecute

SBCommandReturnObjectAppendMessage

SBCommandReturnObjectAppendWarning

SBCommandReturnObjectClear

SBCommandReturnObjectGetDescription

SBCommandReturnObjectGetError

SBCommandReturnObjectGetError2

SBCommandReturnObjectGetErrorSize

SBCommandReturnObjectGetOutput

SBCommandReturnObjectGetOutput2

SBCommandReturnObjectGetOutputSize

SBCommandReturnObjectGetStatus

SBCommandReturnObjectHasResult

SBCommandReturnObjectIsValid

SBCommandReturnObjectPrintf

SBCommandReturnObjectPutCString

SBCommandReturnObjectPutError

SBCommandReturnObjectPutOutput

SBCommandReturnObjectRelease

SBCommandReturnObjectSetError

SBCommandReturnObjectSetError2

SBCommandReturnObjectSetImmediateErrorFile

SBCommandReturnObjectSetImmediateOutputFile

SBCommandReturnObjectSetStatus

SBCommandReturnObjectSucceeded

SBCommandSetHelp

SBCommandSetHelpLong

SBCommunicationAdoptFileDesriptor

SBCommunicationConnect

SBCommunicationDisconnect

SBCommunicationGetBroadcaster

SBCommunicationGetBroadcasterClass

SBCommunicationGetCloseOnEOF

SBCommunicationIsConnected

SBCommunicationIsValid

SBCommunicationRead

SBCommunicationReadThreadIsRunning

SBCommunicationReadThreadStart

SBCommunicationReadThreadStop

SBCommunicationSetCloseOnEOF

SBCommunicationSetReadThreadBytesReceivedCallback

SBCommunicationWrite

SBCompileUnitFindLineEntryIndex

SBCompileUnitFindLineEntryIndex2

SBCompileUnitFindSupportFileIndex

SBCompileUnitGetDescription

SBCompileUnitGetFileSpec

SBCompileUnitGetLanguage

SBCompileUnitGetLineEntryAtIndex

SBCompileUnitGetNumLineEntries

SBCompileUnitGetNumSupportFiles

SBCompileUnitGetSupportFileAtIndex

SBCompileUnitGetTypes

SBCompileUnitIsValid

SBDataAppend

SBDataClear

SBDataCreateDataFromCString

SBDataCreateDataFromDoubleArray

SBDataCreateDataFromSInt32Array

SBDataCreateDataFromSInt64Array

SBDataCreateDataFromUInt32Array

SBDataCreateDataFromUInt64Array

SBDataGetAddress

SBDataGetAddressByteSize

SBDataGetByteOrder

SBDataGetByteSize

SBDataGetDescription

SBDataGetDouble

SBDataGetFloat

SBDataGetLongDouble

SBDataGetSignedInt16

SBDataGetSignedInt32

SBDataGetSignedInt64

SBDataGetSignedInt8

SBDataGetString

SBDataGetUnsignedInt16

SBDataGetUnsignedInt32

SBDataGetUnsignedInt64

SBDataGetUnsignedInt8

SBDataIsValid

SBDataReadRawData

SBDataSetAddressByteSize

SBDataSetByteOrder

SBDataSetData

SBDataSetDataFromCString

SBDataSetDataFromDoubleArray

SBDataSetDataFromSInt32Array

SBDataSetDataFromSInt64Array

SBDataSetDataFromUInt32Array

SBDataSetDataFromUInt64Array

SBDebuggerClear

SBDebuggerCreate

SBDebuggerCreate2

SBDebuggerCreateCategory

SBDebuggerCreateTarget

SBDebuggerCreateTarget2

SBDebuggerCreateTargetWithFileAndArch

SBDebuggerCreateTargetWithFileAndTargetTriple

SBDebuggerDeleteCategory

SBDebuggerDeleteTarget

SBDebuggerDestroy

SBDebuggerDispatchInput

SBDebuggerDispatchInput2

SBDebuggerDispatchInputEndOfFile

SBDebuggerDispatchInputInterrupt

SBDebuggerEnableLog

SBDebuggerFindDebuggerWithID

SBDebuggerFindTargetWithFileAndArch

SBDebuggerFindTargetWithProcessID

SBDebuggerGetAsync

SBDebuggerGetCategory

SBDebuggerGetCategoryAtIndex

SBDebuggerGetCloseInputOnEOF

SBDebuggerGetCommandInterpreter

SBDebuggerGetDefaultArchitecture

SBDebuggerGetDefaultCategory

SBDebuggerGetDescription

SBDebuggerGetErrorFileHandle

SBDebuggerGetFilterForType

SBDebuggerGetFormatForType

SBDebuggerGetID

SBDebuggerGetIndexOfTarget

SBDebuggerGetInputFileHandle

SBDebuggerGetInstanceName

SBDebuggerGetInternalVariableValue

SBDebuggerGetListener

SBDebuggerGetNumCategories

SBDebuggerGetNumTargets

SBDebuggerGetOutputFileHandle

SBDebuggerGetPrompt

SBDebuggerGetScriptLanguage

SBDebuggerGetScriptingLanguage

SBDebuggerGetSelectedPlatform

SBDebuggerGetSelectedTarget

SBDebuggerGetSourceManager

SBDebuggerGetSummaryForType

SBDebuggerGetSyntheticForType

SBDebuggerGetTargetAtIndex

SBDebuggerGetTerminalWidth

SBDebuggerGetUseColor

SBDebuggerGetUseExternalEditor

SBDebuggerGetVersionString

SBDebuggerHandleCommand

SBDebuggerHandleProcessEvent

SBDebuggerInitialize

SBDebuggerIsValid

SBDebuggerMemoryPressureDetected

SBDebuggerPushInputReader

SBDebuggerRestoreInputTerminalState

SBDebuggerRunCommandInterpreter

SBDebuggerRunCommandInterpreter2

SBDebuggerSaveInputTerminalState

SBDebuggerSetAsync

SBDebuggerSetCloseInputOnEOF

SBDebuggerSetCurrentPlatform

SBDebuggerSetCurrentPlatformSDKRoot

SBDebuggerSetDefaultArchitecture

SBDebuggerSetErrorFileHandle

SBDebuggerSetInputFileHandle

SBDebuggerSetInternalVariable

SBDebuggerSetOutputFileHandle

SBDebuggerSetPrompt

SBDebuggerSetScriptLanguage

SBDebuggerSetSelectedPlatform

SBDebuggerSetSelectedTarget

SBDebuggerSetTerminalWidth

SBDebuggerSetUseColor

SBDebuggerSetUseExternalEditor

SBDebuggerSkipAppInitFiles

SBDebuggerSkipLLDBInitFiles

SBDebuggerStateAsCString

SBDebuggerStateIsRunningState

SBDebuggerStateIsStoppedState

SBDebuggerTerminate

SBDeclarationGetColumn

SBDeclarationGetDescription

SBDeclarationGetFileSpec

SBDeclarationGetLine

SBDeclarationIsValid

SBDeclarationSetColumn

SBDeclarationSetFileSpec

SBDeclarationSetLine

SBErrorClear

SBErrorFail

SBErrorGetCString

SBErrorGetDescription

SBErrorGetError

SBErrorGetType

SBErrorIsValid

SBErrorSetError

SBErrorSetErrorString

SBErrorSetErrorStringWithFormat

SBErrorSetErrorToErrno

SBErrorSetErrorToGenericError

SBErrorSuccess

SBEventBroadcasterMatchesPtr

SBEventBroadcasterMatchesRef

SBEventClear

SBEventGetBroadcaster

SBEventGetBroadcasterClass

SBEventGetCStringFromEvent

SBEventGetDataFlavor

SBEventGetDescription

SBEventGetType

SBEventIsValid

SBExecutionContextGetFrame

SBExecutionContextGetProcess

SBExecutionContextGetTarget

SBExecutionContextGetThread

SBExpressionOptionsGetCoerceResultToId

SBExpressionOptionsGetFetchDynamicValue

SBExpressionOptionsGetGenerateDebugInfo

SBExpressionOptionsGetIgnoreBreakpoints

SBExpressionOptionsGetOneThreadTimeoutInMicroSeconds

SBExpressionOptionsGetStopOthers

SBExpressionOptionsGetSuppressPersistentResult

SBExpressionOptionsGetTimeoutInMicroSeconds

SBExpressionOptionsGetTrapExceptions

SBExpressionOptionsGetTryAllThreads

SBExpressionOptionsGetUnwindOnError

SBExpressionOptionsSetCoerceResultToId

SBExpressionOptionsSetFetchDynamicValue

SBExpressionOptionsSetGenerateDebugInfo

SBExpressionOptionsSetIgnoreBreakpoints

SBExpressionOptionsSetLanguage

SBExpressionOptionsSetOneThreadTimeoutInMicroSeconds

SBExpressionOptionsSetStopOthers

SBExpressionOptionsSetSuppressPersistentResult

SBExpressionOptionsSetTimeoutInMicroSeconds

SBExpressionOptionsSetTrapExceptions

SBExpressionOptionsSetTryAllThreads

SBExpressionOptionsSetUnwindOnError

SBFileSpecExists

SBFileSpecGetDescription

SBFileSpecGetDirectory

SBFileSpecGetFilename

SBFileSpecGetPath

SBFileSpecIsValid

SBFileSpecListAppend

SBFileSpecListAppendIfUnique

SBFileSpecListClear

SBFileSpecListFindFileIndex

SBFileSpecListGetDescription

SBFileSpecListGetFileSpecAtIndex

SBFileSpecListGetSize

SBFileSpecResolveExecutableLocation

SBFileSpecResolvePath

SBFileSpecSetDirectory

SBFileSpecSetFilename

SBFrameClear

SBFrameDisassemble

SBFrameEvaluateExpression

SBFrameEvaluateExpression2

SBFrameEvaluateExpression3

SBFrameEvaluateExpression4

SBFrameFindRegister

SBFrameFindValue

SBFrameFindValue2

SBFrameFindVariable

SBFrameFindVariable2

SBFrameGetBlock

SBFrameGetCFA

SBFrameGetCompileUnit

SBFrameGetDescription

SBFrameGetDisplayFunctionName

SBFrameGetFP

SBFrameGetFrameBlock

SBFrameGetFrameID

SBFrameGetFunction

SBFrameGetFunctionName

SBFrameGetLineEntry

SBFrameGetModule

SBFrameGetPC

SBFrameGetPCAddress

SBFrameGetRegisters

SBFrameGetSP

SBFrameGetSymbol

SBFrameGetSymbolContext

SBFrameGetThread

SBFrameGetValueForVariablePath

SBFrameGetValueForVariablePath2

SBFrameGetVariables

SBFrameGetVariables2

SBFrameGetVariables3

SBFrameIsEqual

SBFrameIsInlined

SBFrameIsValid

SBFrameSetPC

SBFunctionGetBlock

SBFunctionGetDescription

SBFunctionGetDisplayName

SBFunctionGetEndAddress

SBFunctionGetInstructions

SBFunctionGetInstructions2

SBFunctionGetIsOptimized

SBFunctionGetLanguage

SBFunctionGetMangledName

SBFunctionGetName

SBFunctionGetPrologueByteSize

SBFunctionGetStartAddress

SBFunctionGetType

SBFunctionIsValid

SBHostOSGetLLDBPath

SBHostOSGetLLDBPythonPath

SBHostOSGetProgramFileSpec

SBInputReaderIsActive

SBInputReaderSetIsDone

SBInstructionDoesBranch

SBInstructionDumpEmulation

SBInstructionEmulateWithFrame

SBInstructionGetAddress

SBInstructionGetAddressClass

SBInstructionGetByteSize

SBInstructionGetComment

SBInstructionGetData

SBInstructionGetDescription

SBInstructionGetMnemonic

SBInstructionGetOperands

SBInstructionIsValid

SBInstructionListAppendInstruction

SBInstructionListClear

SBInstructionListDumpEmulationForAllInstructions

SBInstructionListGetDescription

SBInstructionListGetInstructionAtIndex

SBInstructionListGetSize

SBInstructionListIsValid

SBInstructionListPrint

SBInstructionPrint

SBInstructionTestEmulation

SBLaunchInfoAddCloseFileAction

SBLaunchInfoAddDuplicateFileAction

SBLaunchInfoAddOpenFileAction

SBLaunchInfoAddSuppressFileAction

SBLaunchInfoClear

SBLaunchInfoGetArgumentAtIndex

SBLaunchInfoGetDetachOnError

SBLaunchInfoGetEnvironmentEntryAtIndex

SBLaunchInfoGetExecutableFile

SBLaunchInfoGetGroupID

SBLaunchInfoGetLaunchEventData

SBLaunchInfoGetLaunchFlags

SBLaunchInfoGetListener

SBLaunchInfoGetNumArguments

SBLaunchInfoGetNumEnvironmentEntries

SBLaunchInfoGetProcessID

SBLaunchInfoGetProcessPluginName

SBLaunchInfoGetResumeCount

SBLaunchInfoGetShell

SBLaunchInfoGetShellExpandArguments

SBLaunchInfoGetUserID

SBLaunchInfoGetWorkingDirectory

SBLaunchInfoGroupIDIsValid

SBLaunchInfoSetArguments

SBLaunchInfoSetDetachOnError

SBLaunchInfoSetEnvironmentEntries

SBLaunchInfoSetExecutableFile

SBLaunchInfoSetGroupID

SBLaunchInfoSetLaunchEventData

SBLaunchInfoSetLaunchFlags

SBLaunchInfoSetListener

SBLaunchInfoSetProcessPluginName

SBLaunchInfoSetResumeCount

SBLaunchInfoSetShell

SBLaunchInfoSetShellExpandArguments

SBLaunchInfoSetUserID

SBLaunchInfoSetWorkingDirectory

SBLaunchInfoUserIDIsValid

SBLineEntryGetColumn

SBLineEntryGetDescription

SBLineEntryGetEndAddress

SBLineEntryGetFileSpec

SBLineEntryGetLine

SBLineEntryGetStartAddress

SBLineEntryIsValid

SBLineEntrySetColumn

SBLineEntrySetFileSpec

SBLineEntrySetLine

SBListenerAddEvent

SBListenerClear

SBListenerGetNextEvent

SBListenerGetNextEventForBroadcaster

SBListenerGetNextEventForBroadcasterWithType

SBListenerHandleBroadcastEvent

SBListenerIsValid

SBListenerPeekAtNextEvent

SBListenerPeekAtNextEventForBroadcaster

SBListenerPeekAtNextEventForBroadcasterWithType

SBListenerStartListeningForEventClass

SBListenerStartListeningForEvents

SBListenerStopListeningForEventClass

SBListenerStopListeningForEvents

SBListenerWaitForEvent

SBListenerWaitForEventForBroadcaster

SBListenerWaitForEventForBroadcasterWithType

SBModuleClear

SBModuleFindFirstGlobalVariable

SBModuleFindFirstType

SBModuleFindFunctions

SBModuleFindGlobalVariables

SBModuleFindSection

SBModuleFindSymbol

SBModuleFindSymbols

SBModuleFindTypes

SBModuleGetAddressByteSize

SBModuleGetBasicType

SBModuleGetByteOrder

SBModuleGetCompileUnitAtIndex

SBModuleGetDescription

SBModuleGetFileSpec

SBModuleGetNumCompileUnits

SBModuleGetNumSections

SBModuleGetNumSymbols

SBModuleGetObjectFileHeaderAddress

SBModuleGetPlatformFileSpec

SBModuleGetRemoteInstallFileSpec

SBModuleGetSectionAtIndex

SBModuleGetSymbolAtIndex

SBModuleGetSymbolFileSpec

SBModuleGetTriple

SBModuleGetTypeByID

SBModuleGetTypes

SBModuleGetUUIDBytes

SBModuleGetUUIDString

SBModuleGetVersion

SBModuleIsValid

SBModuleResolveFileAddress

SBModuleResolveSymbolContextForAddress

SBModuleSetPlatformFileSpec

SBModuleSetRemoteInstallFileSpec

SBModuleSpecClear

SBModuleSpecGetDescription

SBModuleSpecGetFileSpec

SBModuleSpecGetObjectName

SBModuleSpecGetPlatformFileSpec

SBModuleSpecGetSymbolFileSpec

SBModuleSpecGetTriple

SBModuleSpecGetUUIDBytes

SBModuleSpecGetUUIDLength

SBModuleSpecIsValid

SBModuleSpecListAppend

SBModuleSpecListAppend2

SBModuleSpecListFindFirstMatchingSpec

SBModuleSpecListFindMatchingSpecs

SBModuleSpecListGetDescription

SBModuleSpecListGetModuleSpecifications

SBModuleSpecListGetSize

SBModuleSpecListGetSpecAtIndex

SBModuleSpecSetFileSpec

SBModuleSpecSetObjectName

SBModuleSpecSetPlatformFileSpec

SBModuleSpecSetSymbolFileSpec

SBModuleSpecSetTriple

SBModuleSpecSetUUIDBytes

SBPlatformClear

SBPlatformConnectOptionsDisableRsync

SBPlatformConnectOptionsEnableRsync

SBPlatformConnectOptionsGetLocalCacheDirectory

SBPlatformConnectOptionsGetRsyncEnabled

SBPlatformConnectOptionsGetURL

SBPlatformConnectOptionsSetLocalCacheDirectory

SBPlatformConnectOptionsSetURL

SBPlatformConnectRemote

SBPlatformDisconnectRemote

SBPlatformGet

SBPlatformGetFilePermissions

SBPlatformGetHostname

SBPlatformGetName

SBPlatformGetOSBuild

SBPlatformGetOSDescription

SBPlatformGetOSMajorVersion

SBPlatformGetOSMinorVersion

SBPlatformGetOSUpdateVersion

SBPlatformGetTriple

SBPlatformGetWorkingDirectory

SBPlatformInstall

SBPlatformIsConnected

SBPlatformIsValid

SBPlatformKill

SBPlatformLaunch

SBPlatformMakeDirectory

SBPlatformPut

SBPlatformRun

SBPlatformSetFilePermissions

SBPlatformSetWorkingDirectory

SBPlatformShellCommandClear

SBPlatformShellCommandGetCommand

SBPlatformShellCommandGetOutput

SBPlatformShellCommandGetSignal

SBPlatformShellCommandGetStatus

SBPlatformShellCommandGetTimeoutSeconds

SBPlatformShellCommandGetWorkingDirectory

SBPlatformShellCommandSetCommand

SBPlatformShellCommandSetTimeoutSeconds

SBPlatformShellCommandSetWorkingDirectory

SBProcessAppendEventStateReport

SBProcessClear

SBProcessContinue

SBProcessCreateOSPluginThread

SBProcessDestroy

SBProcessDetach

SBProcessDetach2

SBProcessEventIsProcessEvent

SBProcessGetAddressByteSize

SBProcessGetAsyncProfileData

SBProcessGetBroadcaster

SBProcessGetBroadcasterClass

SBProcessGetBroadcasterClassName

SBProcessGetByteOrder

SBProcessGetDescription

SBProcessGetExitDescription

SBProcessGetExitStatus

SBProcessGetExtendedBacktraceTypeAtIndex

SBProcessGetHistoryThreads

SBProcessGetNumExtendedBacktraceTypes

SBProcessGetNumQueues

SBProcessGetNumRestartedReasonsFromEvent

SBProcessGetNumSupportedHardwareWatchpoints

SBProcessGetNumThreads

SBProcessGetPluginName

SBProcessGetProcessFromEvent

SBProcessGetProcessID

SBProcessGetQueueAtIndex

SBProcessGetRestartedFromEvent

SBProcessGetRestartedReasonAtIndexFromEvent

SBProcessGetSTDERR

SBProcessGetSTDOUT

SBProcessGetSelectedThread

SBProcessGetShortPluginName

SBProcessGetState

SBProcessGetStateFromEvent

SBProcessGetStopID

SBProcessGetTarget

SBProcessGetThreadAtIndex

SBProcessGetThreadByID

SBProcessGetThreadByIndexID

SBProcessGetUniqueID

SBProcessGetUnixSignals

SBProcessIsInstrumentationRuntimePresent

SBProcessIsValid

SBProcessKill

SBProcessLoadImage

SBProcessPutSTDIN

SBProcessReadCStringFromMemory

SBProcessReadMemory

SBProcessReadPointerFromMemory

SBProcessReadUnsignedFromMemory

SBProcessRemoteAttachToProcessWithID

SBProcessRemoteLaunch

SBProcessReportEventState

SBProcessSendAsyncInterrupt

SBProcessSendEventData

SBProcessSetSelectedThread

SBProcessSetSelectedThreadByID

SBProcessSetSelectedThreadByIndexID

SBProcessSignal

SBProcessStop

SBProcessUnloadImage

SBProcessWriteMemory

SBQueueClear

SBQueueGetIndexID

SBQueueGetKind

SBQueueGetName

SBQueueGetNumPendingItems

SBQueueGetNumRunningItems

SBQueueGetNumThreads

SBQueueGetPendingItemAtIndex

SBQueueGetProcess

SBQueueGetQueueID

SBQueueGetThreadAtIndex

SBQueueIsValid

SBQueueItemClear

SBQueueItemGetAddress

SBQueueItemGetExtendedBacktraceThread

SBQueueItemGetKind

SBQueueItemIsValid

SBQueueItemSetAddress

SBQueueItemSetKind

SBSectionFindSubSection

SBSectionGetByteSize

SBSectionGetDescription

SBSectionGetFileAddress

SBSectionGetFileByteSize

SBSectionGetFileOffset

SBSectionGetLoadAddress

SBSectionGetName

SBSectionGetNumSubSections

SBSectionGetParent

SBSectionGetSectionData

SBSectionGetSectionData2

SBSectionGetSectionType

SBSectionGetSubSectionAtIndex

SBSectionGetTargetByteSize

SBSectionIsValid

SBSourceManagerDisplaySourceLinesWithLineNumbers

SBStreamClear

SBStreamGetData

SBStreamGetSize

SBStreamIsValid

SBStreamPrintf

SBStreamRedirectToFile

SBStreamRedirectToFileDescriptor

SBStreamRedirectToFileHandle

SBStringListAppendList

SBStringListAppendList2

SBStringListAppendString

SBStringListClear

SBStringListGetSize

SBStringListGetStringAtIndex

SBStringListIsValid

SBSymbolContextGetBlock

SBSymbolContextGetCompileUnit

SBSymbolContextGetDescription

SBSymbolContextGetFunction

SBSymbolContextGetLineEntry

SBSymbolContextGetModule

SBSymbolContextGetParentOfInlinedScope

SBSymbolContextGetSymbol

SBSymbolContextIsValid

SBSymbolContextListAppend

SBSymbolContextListAppend2

SBSymbolContextListClear

SBSymbolContextListGetContextAtIndex

SBSymbolContextListGetDescription

SBSymbolContextListGetSize

SBSymbolContextListIsValid

SBSymbolContextSetBlock

SBSymbolContextSetCompileUnit

SBSymbolContextSetFunction

SBSymbolContextSetLineEntry

SBSymbolContextSetModule

SBSymbolContextSetSymbol

SBSymbolGetDescription

SBSymbolGetDisplayName

SBSymbolGetEndAddress

SBSymbolGetInstructions

SBSymbolGetInstructions2

SBSymbolGetMangledName

SBSymbolGetName

SBSymbolGetPrologueByteSize

SBSymbolGetStartAddress

SBSymbolGetType

SBSymbolIsExternal

SBSymbolIsSynthetic

SBSymbolIsValid

SBTargetAddModule

SBTargetAddModule2

SBTargetAddModule3

SBTargetAddModule4

SBTargetAttach

SBTargetAttachToProcessWithID

SBTargetAttachToProcessWithName

SBTargetBreakpointCreateByAddress

SBTargetBreakpointCreateByLocation

SBTargetBreakpointCreateByLocation2

SBTargetBreakpointCreateByName

SBTargetBreakpointCreateByName2

SBTargetBreakpointCreateByName3

SBTargetBreakpointCreateByNames

SBTargetBreakpointCreateByRegex

SBTargetBreakpointCreateByRegex2

SBTargetBreakpointCreateBySourceRegex

SBTargetBreakpointCreateBySourceRegex2

SBTargetBreakpointCreateForException

SBTargetBreakpointDelete

SBTargetClear

SBTargetClearModuleLoadAddress

SBTargetClearSectionLoadAddress

SBTargetConnectRemote

SBTargetCreateValueFromAddress

SBTargetCreateValueFromData

SBTargetCreateValueFromExpression

SBTargetDeleteAllBreakpoints

SBTargetDeleteAllWatchpoints

SBTargetDeleteWatchpoint

SBTargetDisableAllBreakpoints

SBTargetDisableAllWatchpoints

SBTargetEnableAllBreakpoints

SBTargetEnableAllWatchpoints

SBTargetEvaluateExpression

SBTargetEventIsTargetEvent

SBTargetFindBreakpointByID

SBTargetFindFirstGlobalVariable

SBTargetFindFirstType

SBTargetFindFunctions

SBTargetFindGlobalFunctions

SBTargetFindGlobalVariables

SBTargetFindGlobalVariables2

SBTargetFindModule

SBTargetFindSymbols

SBTargetFindTypes

SBTargetFindWatchpointByID

SBTargetGetAddressByteSize

SBTargetGetBasicType

SBTargetGetBreakpointAtIndex

SBTargetGetBroadcaster

SBTargetGetBroadcasterClassName

SBTargetGetByteOrder

SBTargetGetCodeByteSize

SBTargetGetDataByteSize

SBTargetGetDebugger

SBTargetGetDescription

SBTargetGetExecutable

SBTargetGetInstructions

SBTargetGetInstructions2

SBTargetGetInstructionsWithFlavor

SBTargetGetInstructionsWithFlavor2

SBTargetGetLaunchInfo

SBTargetGetModuleAtIndex

SBTargetGetModuleAtIndexFromEvent

SBTargetGetNumBreakpoints

SBTargetGetNumModules

SBTargetGetNumModulesFromEvent

SBTargetGetNumWatchpoints

SBTargetGetPlatform

SBTargetGetProcess

SBTargetGetSourceManager

SBTargetGetStackRedZoneSize

SBTargetGetTargetFromEvent

SBTargetGetTriple

SBTargetGetWatchpointAtIndex

SBTargetInstall

SBTargetIsValid

SBTargetLaunch

SBTargetLaunch2

SBTargetLaunchSimple

SBTargetLoadCore

SBTargetReadInstructions

SBTargetReadInstructions2

SBTargetReadMemory

SBTargetRemoveModule

SBTargetResolveFileAddress

SBTargetResolveLoadAddress

SBTargetResolvePastLoadAddress

SBTargetResolveSymbolContextForAddress

SBTargetSetLaunchInfo

SBTargetSetModuleLoadAddress

SBTargetSetSectionLoadAddress

SBTargetWatchAddress

SBThreadClear

SBThreadCollectionGetSize

SBThreadCollectionGetThreadAtIndex

SBThreadCollectionIsValid

SBThreadEventIsThreadEvent

SBThreadGetBroadcasterClassName

SBThreadGetDescription

SBThreadGetExtendedBacktraceOriginatingIndexID

SBThreadGetExtendedBacktraceThread

SBThreadGetFrameAtIndex

SBThreadGetIndexID

SBThreadGetInfoItemByPathAsString

SBThreadGetName

SBThreadGetNumFrames

SBThreadGetProcess

SBThreadGetQueue

SBThreadGetQueueID

SBThreadGetQueueName

SBThreadGetSelectedFrame

SBThreadGetStackFrameFromEvent

SBThreadGetStatus

SBThreadGetStopDescription

SBThreadGetStopReason

SBThreadGetStopReasonDataAtIndex

SBThreadGetStopReasonDataCount

SBThreadGetStopReasonExtendedInfoAsJSON

SBThreadGetStopReturnValue

SBThreadGetThreadFromEvent

SBThreadGetThreadID

SBThreadIsStopped

SBThreadIsSuspended

SBThreadIsValid

SBThreadJumpToLine

SBThreadPlanClear

SBThreadPlanGetDescription

SBThreadPlanGetStopReason

SBThreadPlanGetStopReasonDataAtIndex

SBThreadPlanGetStopReasonDataCount

SBThreadPlanGetThread

SBThreadPlanIsPlanComplete

SBThreadPlanIsValid

SBThreadPlanQueueThreadPlanForRunToAddress

SBThreadPlanQueueThreadPlanForStepInRange

SBThreadPlanQueueThreadPlanForStepOut

SBThreadPlanQueueThreadPlanForStepOverRange

SBThreadPlanSetPlanComplete

SBThreadResume

SBThreadReturnFromFrame

SBThreadRunToAddress

SBThreadSafeToCallFunctions

SBThreadSetSelectedFrame

SBThreadStepInstruction

SBThreadStepInto

SBThreadStepInto2

SBThreadStepOut

SBThreadStepOutOfFrame

SBThreadStepOver

SBThreadStepOverUntil

SBThreadStepUsingScriptedThreadPlan

SBThreadSuspend

SBTypeCategoryAddTypeFilter

SBTypeCategoryAddTypeFormat

SBTypeCategoryAddTypeSummary

SBTypeCategoryAddTypeSynthetic

SBTypeCategoryDeleteTypeFilter

SBTypeCategoryDeleteTypeFormat

SBTypeCategoryDeleteTypeSummary

SBTypeCategoryDeleteTypeSynthetic

SBTypeCategoryGetDescription

SBTypeCategoryGetEnabled

SBTypeCategoryGetFilterAtIndex

SBTypeCategoryGetFilterForType

SBTypeCategoryGetFormatAtIndex

SBTypeCategoryGetFormatForType

SBTypeCategoryGetName

SBTypeCategoryGetNumFilters

SBTypeCategoryGetNumFormats

SBTypeCategoryGetNumSummaries

SBTypeCategoryGetNumSynthetics

SBTypeCategoryGetSummaryAtIndex

SBTypeCategoryGetSummaryForType

SBTypeCategoryGetSyntheticAtIndex

SBTypeCategoryGetSyntheticForType

SBTypeCategoryGetTypeNameSpecifierForFilterAtIndex

SBTypeCategoryGetTypeNameSpecifierForFormatAtIndex

SBTypeCategoryGetTypeNameSpecifierForSummaryAtIndex

SBTypeCategoryGetTypeNameSpecifierForSyntheticAtIndex

SBTypeCategoryIsValid

SBTypeCategorySetEnabled

SBTypeEnumMemberGetDescription

SBTypeEnumMemberGetName

SBTypeEnumMemberGetType

SBTypeEnumMemberGetValueAsSigned

SBTypeEnumMemberGetValueAsUnsigned

SBTypeEnumMemberIsValid

SBTypeEnumMemberListAppend

SBTypeEnumMemberListGetSize

SBTypeEnumMemberListGetTypeEnumMemberAtIndex

SBTypeEnumMemberListIsValid

SBTypeFilterAppendExpressionPath

SBTypeFilterClear

SBTypeFilterGetDescription

SBTypeFilterGetExpressionPathAtIndex

SBTypeFilterGetNumberOfExpressionPaths

SBTypeFilterGetOptions

SBTypeFilterIsEqualTo

SBTypeFilterIsValid

SBTypeFilterReplaceExpressionPathAtIndex

SBTypeFilterSetOptions

SBTypeFormatGetDescription

SBTypeFormatGetFormat

SBTypeFormatGetOptions

SBTypeFormatGetTypeName

SBTypeFormatIsEqualTo

SBTypeFormatIsValid

SBTypeFormatSetFormat

SBTypeFormatSetOptions

SBTypeFormatSetTypeName

SBTypeGetArrayElementType

SBTypeGetBasicType

SBTypeGetBasicType2

SBTypeGetByteSize

SBTypeGetCanonicalType

SBTypeGetDereferencedType

SBTypeGetDescription

SBTypeGetDirectBaseClassAtIndex

SBTypeGetDisplayTypeName

SBTypeGetEnumMembers

SBTypeGetFieldAtIndex

SBTypeGetFunctionArgumentTypes

SBTypeGetFunctionReturnType

SBTypeGetMemberFunctionAtIndex

SBTypeGetName

SBTypeGetNumberOfDirectBaseClasses

SBTypeGetNumberOfFields

SBTypeGetNumberOfMemberFunctions

SBTypeGetNumberOfTemplateArguments

SBTypeGetNumberOfVirtualBaseClasses

SBTypeGetPointeeType

SBTypeGetPointerType

SBTypeGetReferenceType

SBTypeGetTemplateArgumentKind

SBTypeGetTemplateArgumentType

SBTypeGetTypeClass

SBTypeGetTypeFlags

SBTypeGetTypedefedType

SBTypeGetUnqualifiedType

SBTypeGetVectorElementType

SBTypeGetVirtualBaseClassAtIndex

SBTypeIsArrayType

SBTypeIsFunctionType

SBTypeIsPointerType

SBTypeIsPolymorphicClass

SBTypeIsReferenceType

SBTypeIsTypeComplete

SBTypeIsTypedefType

SBTypeIsValid

SBTypeIsVectorType

SBTypeListAppend

SBTypeListGetSize

SBTypeListGetTypeAtIndex

SBTypeListIsValid

SBTypeMemberFunctionGetArgumentTypeAtIndex

SBTypeMemberFunctionGetDescription

SBTypeMemberFunctionGetKind

SBTypeMemberFunctionGetName

SBTypeMemberFunctionGetNumberOfArguments

SBTypeMemberFunctionGetReturnType

SBTypeMemberFunctionGetType

SBTypeMemberFunctionIsValid

SBTypeMemberGetBitfieldSizeInBits

SBTypeMemberGetDescription

SBTypeMemberGetName

SBTypeMemberGetOffsetInBits

SBTypeMemberGetOffsetInBytes

SBTypeMemberGetType

SBTypeMemberIsBitfield

SBTypeMemberIsValid

SBTypeNameSpecifierGetDescription

SBTypeNameSpecifierGetName

SBTypeNameSpecifierGetType

SBTypeNameSpecifierIsEqualTo

SBTypeNameSpecifierIsRegex

SBTypeNameSpecifierIsValid

SBTypeSummaryCreateWithFunctionName

SBTypeSummaryCreateWithScriptCode

SBTypeSummaryCreateWithSummaryString

SBTypeSummaryGetData

SBTypeSummaryGetDescription

SBTypeSummaryGetOptions

SBTypeSummaryIsEqualTo

SBTypeSummaryIsFunctionCode

SBTypeSummaryIsFunctionName

SBTypeSummaryIsSummaryString

SBTypeSummaryIsValid

SBTypeSummaryOptionsGetCapping

SBTypeSummaryOptionsGetLanguage

SBTypeSummaryOptionsIsValid

SBTypeSummaryOptionsSetCapping

SBTypeSummaryOptionsSetLanguage

SBTypeSummarySetFunctionCode

SBTypeSummarySetFunctionName

SBTypeSummarySetOptions

SBTypeSummarySetSummaryString

SBTypeSyntheticCreateWithClassName

SBTypeSyntheticCreateWithScriptCode

SBTypeSyntheticGetData

SBTypeSyntheticGetDescription

SBTypeSyntheticGetOptions

SBTypeSyntheticIsClassCode

SBTypeSyntheticIsClassName

SBTypeSyntheticIsEqualTo

SBTypeSyntheticIsValid

SBTypeSyntheticSetClassCode

SBTypeSyntheticSetClassName

SBTypeSyntheticSetOptions

SBUnixSignalsClear

SBUnixSignalsGetNumSignals

SBUnixSignalsGetShouldNotify

SBUnixSignalsGetShouldStop

SBUnixSignalsGetShouldSuppress

SBUnixSignalsGetSignalAsCString

SBUnixSignalsGetSignalAtIndex

SBUnixSignalsGetSignalNumberFromName

SBUnixSignalsIsValid

SBUnixSignalsSetShouldNotify

SBUnixSignalsSetShouldStop

SBUnixSignalsSetShouldSuppress

SBValueAddressOf

SBValueCast

SBValueClear

SBValueCreateChildAtOffset

SBValueCreateValueFromAddress

SBValueCreateValueFromData

SBValueCreateValueFromExpression

SBValueCreateValueFromExpression2

SBValueDereference

SBValueGetAddress

SBValueGetByteSize

SBValueGetChildAtIndex

SBValueGetChildAtIndex2

SBValueGetChildMemberWithName

SBValueGetChildMemberWithName2

SBValueGetData

SBValueGetDeclaration

SBValueGetDescription

SBValueGetDisplayTypeName

SBValueGetDynamicValue

SBValueGetError

SBValueGetExpressionPath

SBValueGetExpressionPath2

SBValueGetFormat

SBValueGetFrame

SBValueGetID

SBValueGetIndexOfChildWithName

SBValueGetLoadAddress

SBValueGetLocation

SBValueGetName

SBValueGetNonSyntheticValue

SBValueGetNumChildren

SBValueGetObjectDescription

SBValueGetOpaqueType

SBValueGetPointeeData

SBValueGetPreferDynamicValue

SBValueGetPreferSyntheticValue

SBValueGetProcess

SBValueGetStaticValue

SBValueGetSummary

SBValueGetSummary2

SBValueGetTarget

SBValueGetThread

SBValueGetType

SBValueGetTypeFilter

SBValueGetTypeFormat

SBValueGetTypeName

SBValueGetTypeSummary

SBValueGetTypeSynthetic

SBValueGetTypeValidatorResult

SBValueGetValue

SBValueGetValueAsSigned

SBValueGetValueAsSigned2

SBValueGetValueAsUnsigned

SBValueGetValueAsUnsigned2

SBValueGetValueDidChange

SBValueGetValueForExpressionPath

SBValueGetValueType

SBValueIsDynamic

SBValueIsInScope

SBValueIsRuntimeSupportValue

SBValueIsSynthetic

SBValueIsValid

SBValueListAppend

SBValueListAppend2

SBValueListClear

SBValueListFindValueObjectByUID

SBValueListGetFirstValueByName

SBValueListGetSize

SBValueListGetValueAtIndex

SBValueListIsValid

SBValueMightHaveChildren

SBValuePersist

SBValueSetData

SBValueSetFormat

SBValueSetPreferDynamicValue

SBValueSetPreferSyntheticValue

SBValueSetValueFromCString

SBValueSetValueFromCString2

SBValueTypeIsPointerType

SBValueWatch

SBValueWatch2

SBValueWatchPointee

SBVariablesOptionsGetInScopeOnly

SBVariablesOptionsGetIncludeArguments

SBVariablesOptionsGetIncludeLocals

SBVariablesOptionsGetIncludeRuntimeSupportValues

SBVariablesOptionsGetIncludeStatics

SBVariablesOptionsGetUseDynamic

SBVariablesOptionsIsValid

SBVariablesOptionsSetInScopeOnly

SBVariablesOptionsSetIncludeArguments

SBVariablesOptionsSetIncludeLocals

SBVariablesOptionsSetIncludeRuntimeSupportValues

SBVariablesOptionsSetIncludeStatics

SBVariablesOptionsSetUseDynamic

SBWatchpointClear

SBWatchpointEventIsWatchpointEvent

SBWatchpointGetCondition

SBWatchpointGetDescription

SBWatchpointGetError

SBWatchpointGetHardwareIndex

SBWatchpointGetHitCount

SBWatchpointGetID

SBWatchpointGetIgnoreCount

SBWatchpointGetWatchAddress

SBWatchpointGetWatchSize

SBWatchpointGetWatchpointEventTypeFromEvent

SBWatchpointGetWatchpointFromEvent

SBWatchpointIsEnabled

SBWatchpointIsValid

SBWatchpointSetCondition

SBWatchpointSetEnabled

SBWatchpointSetIgnoreCount

Type Definitions

ReadThreadBytesReceived

SBAddressRef

SBAttachInfoRef

SBBlockRef

SBBreakpointLocationRef

SBBreakpointRef

SBBroadcasterRef

SBCommandInterpreterRef

SBCommandInterpreterRunOptionsRef

SBCommandPluginInterfaceRef

SBCommandRef

SBCommandReturnObjectRef

SBCommunicationRef

SBCompileUnitRef

SBDataRef

SBDebuggerRef

SBDeclarationRef

SBErrorRef

SBEventListRef

SBEventRef

SBExecutionContextRef

SBExpressionOptionsRef

SBFileSpecListRef

SBFileSpecRef

SBFrameRef

SBFunctionRef

SBHostOSRef

SBInputReaderRef

SBInstructionListRef

SBInstructionRef

SBLaunchInfoRef

SBLineEntryRef

SBListenerRef

SBModuleRef

SBModuleSpecListRef

SBModuleSpecRef

SBPlatformConnectOptionsRef

SBPlatformRef

SBPlatformShellCommandRef

SBProcessRef

SBQueueItemRef

SBQueueRef

SBSectionRef

SBSourceManagerRef

SBStreamRef

SBStringListRef

SBSymbolContextListRef

SBSymbolContextRef

SBSymbolRef

SBTargetRef

SBThreadCollectionRef

SBThreadPlanRef

SBThreadRef

SBTypeCategoryRef

SBTypeEnumMemberListRef

SBTypeEnumMemberRef

SBTypeFilterRef

SBTypeFormatRef

SBTypeListRef

SBTypeMemberFunctionRef

SBTypeMemberRef

SBTypeNameSpecifierRef

SBTypeRef

SBTypeSummaryOptionsRef

SBTypeSummaryRef

SBTypeSyntheticRef

SBUnixSignalsRef

SBValueListRef

SBValueRef

SBVariablesOptionsRef

SBWatchpointRef

lldb_addr_t
Storage for the value of an address.

lldb_offset_t
Storage for an offset between 2 addresses in memory.

lldb_pid_t
Storage for an OS process ID.

lldb_tid_t
Storage for an OS thread ID.

lldb_user_id_t
Storage for an OS user ID.